consolidate grid data functions
